Move to "submodule" system
Instead of README and scripting, use git submodules to fetch the designated
repositories in a reproducible way. This repository should be named something
like "topic-desktop-platform" now, since it's not a "meta" layer.

Adjust scripts and README to match the new situation.

Fix the layer.conf to properly refer to itself.