Use default device manager
Apparently the desktop is tied to "udev", so make it happy and don't use mdev.
Using x11-base in the image already starts the X server, so remove xfce-selector.
Remove the xserver config, it only breaks things.
Only issue now is to (really) assign "fb0" as default to the LCD display, so
the X server picks that instead of the external HDMI (even when not attached).