qt4-x11-free: Apply GCC8 fix to qt4-x11 as well
The fix to remove "volatile" also applies to the x11 version of Qt, so apply the
patch for both variants by moving it to the common qt4-${PV} include file. This fixes
that the 32-bit version of qt4-x11 fails to build with the same error as qt4-embedded.

Upstream-status: Pending